Backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pypdf
openai

python-dotenv
